kernel: Fix building of extended.cl
Fixes #3132

+++ data/kernels/extended.cl
@@ -275,7 +275,7 @@ encrypt_tea(unsigned int *arg)
 float
 tpdf(unsigned int urandom)
 {
-  float frandom = (float)urandom / 0xFFFFFFFFu;
+  float frandom = (float)urandom / (float)0xFFFFFFFFu;
 
   return (frandom < 0.5f ? (sqrt(2.0f*frandom) - 1.0f) : (1.0f - sqrt(2.0f*(1.0f - frandom))));
 }
